requirements.txt
amqp==5.3.1
anyio==4.12.1
asgiref==3.11.1
billiard==4.1.0
celery==5.3.4
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.5
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
crispy-bootstrap5==2023.10
cron_descriptor==2.0.6
cryptography==46.0.5
csscompressor==0.9.5
defusedxml==0.7.1
Django==4.2.29
django-allauth==0.58.2
django-anymail==10.2
django-appconf==1.2.0
django-axes==6.3.0
django-celery-beat==2.5.0
django-celery-results==2.5.1
django-compressor==4.4
django-cors-headers==4.3.1
django-crispy-forms==2.1
django-environ==0.13.0
django-ratelimit==4.1.0
django-redis==5.4.0
django-timezone-field==7.2.1
djangorestframework==3.14.0
drf-yasg==1.21.7
geographiclib==2.1
geopy==2.4.1
gunicorn==21.2.0
h11==0.16.0
httpcore==1.0.9
httpx==0.25.1
idna==3.11
inflection==0.5.1
jsmin==3.0.1
kombu==5.3.4
numpy==2.4.3
oauthlib==3.3.1
packaging==26.0
pandas==3.0.1
prompt_toolkit==3.0.52
pycparser==3.0
PyJWT==2.11.0
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-json-logger==2.0.7
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0.1
rcssmin==1.1.1
redis==7.3.0
requests==2.32.5
requests-oauthlib==2.0.0
rjsmin==1.2.1
setuptools==69.5.1
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.5
typing_extensions==4.15.0
tzdata==2025.3
uritemplate==4.2.0
urllib3==2.6.3
vine==5.1.0
wcwidth==0.6.0
wheel==0.46.3
whitenoise==6.6.0
